+++ scripts/restart_sctp.sh
@@ -1,3 +1,0 @@
-export LD_LIBRARY_PATH=../sc-machine/bin
-./build_kb.sh
-../sc-machine/bin/sctp-server ../config/sc-web.ini