+++ scripts/run_scweb.sh
@@ -1,5 +1,2 @@
 #!/bin/bash
-
-cd ../sc-web/server/
-python3 app.py
-cd -
+../sc-web/scripts/run_scweb.sh